+++ packages/lila-webpack-config/src/build.js
@@ -1,44 +1,20 @@
 import path from 'path';
-import forEach from 'lodash/forEach';
 import MiniCssExtractPlugin from 'mini-css-extract-plugin';
 import OptimizeCssAssetsPlugin from 'optimize-css-assets-webpack-plugin';
-import HtmlWebpackIncludeAssetsPlugin from 'html-webpack-include-assets-plugin';
 import WebpackBar from 'webpackbar';
 import SpeedMeasurePlugin from 'speed-measure-webpack-plugin';
-import dll from './dll';
 import base from './base';
 import { defaultEntry } from '../../../util/constants';
 
 const { join } = path;
 
 export default (lila, webpack, { entry, cmd, config }) => {
-  const { DllReferencePlugin } = webpack;
   const { getSettings } = lila;
-  const [root, srcDir, buildDir, tmpDir] = getSettings([
-    'root',
-    'src',
-    'build',
-    'tmp',
-  ]);
+  const [root, srcDir, buildDir] = getSettings(['root', 'src', 'build']);
   const srcPath = join(root, srcDir);
   const buildPath = join(root, buildDir);
 
-  const { staticServer = '', minCss = !0, splitJs = {} } = config;
-
-  const dllConfigs = [];
-  const dllPlugins = [];
-  forEach(splitJs, (value, key) => {
-    dllConfigs.push(dll(lila, webpack, { entry, cmd, config }, key, value));
-    dllPlugins.push(
-      new DllReferencePlugin({
-        manifest: join(
-          root,
-          tmpDir,
-          `dll/${entry === defaultEntry ? '' : `${entry}/`}${key}.json`
-        ),
-      })
-    );
-  });
+  const { staticServer = '', minCss = !0 } = config;
 
   const baseConfig = base(lila, webpack, { entry, cmd, config });
 
@@ -48,11 +24,6 @@ export default (lila, webpack, { entry, cmd, config }) => {
     // css standalone
     new MiniCssExtractPlugin({
       filename: '[chunkhash].css',
-    }),
-    // insert split chunk js to html
-    new HtmlWebpackIncludeAssetsPlugin({
-      assets: [{ path: 'dll', glob: '*.js', globPath: `${buildPath}/dll/` }],
-      append: false,
     })
   );
 
@@ -65,9 +36,7 @@ export default (lila, webpack, { entry, cmd, config }) => {
       })
     );
 
-  if (dllPlugins.length) baseConfig.plugins.push(...dllPlugins);
-
-  const buildConfig = {
+  return {
     entry: `${srcPath}/${entry === defaultEntry ? '' : `${entry}/`}index.js`,
     output: {
       path: buildPath,
@@ -77,6 +46,4 @@ export default (lila, webpack, { entry, cmd, config }) => {
     },
     ...baseConfig,
   };
-
-  return [...dllConfigs, buildConfig];
 };