+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.6 [2018-12-13]
 
 - Fix bug of `/` can get mock data.
+- `lila-webpack-config` remove `splitJs` config option.
 
 ## 0.5.5 [2018-12-11]